Move landing-page related elements to Landing.js. Move SubdomainRoute and PrivateRoute consts to App.js. Wrap Component argument with corresponding parent wrapper element.
