+++ client/locale/en.json
@@ -46,6 +46,7 @@
       "name_required": "Name is required",
       "email_required": "Email is required",
       "message_body_required": "Message is required",
+      "wrong_email_format": "Wrong email format",
       "flash": {
         "success_create_contact_message": "Thank you for your message. We will contact you back soon!",
         "error_create_contact_message": "Something went wrong when sending your message. Try sending us again."
