+++ client/src/components/Layouts/HeaderNav.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { Link, Route } from 'react-router-dom'
+import { Link } from 'react-router-dom'
 import decode from 'jwt-decode'
 import { Image, Dropdown, Menu, Label, Icon } from 'semantic-ui-react'
 import { graphql, compose } from 'react-apollo'
@@ -15,43 +15,7 @@ import T from 'i18n-react'
 import logoInverted from '../../images/logo-inverted.png'
 import avatarPlaceholderSmall from '../../images/avatar-placeholder-small.png'
 
-const ActiveLink = ({ label, to, activeOnlyWhenExact }) => (
-  <Route path={to} exact={activeOnlyWhenExact} children={({ match }) => (
-    <Link className={match ? 'active item' : 'item' } to={to}>{label}</Link>
-  )} />
-)
-
-/* jQuery */
-import $ from 'jquery'
-$.animate = require('jquery.easing')
-$.fn.transition = require('semantic-ui-transition')
-$.fn.visibility = require('semantic-ui-visibility')
-
 class HeaderNav extends Component {
-
-  componentDidMount = () => {     
-    // fix menu when passed
-    $('.masthead .ui.text.container')
-    .visibility({
-      once: false,
-      onBottomPassed: function()  {
-        $('.fixed.menu').transition('fade in')
-      },
-      onBottomPassedReverse: function()  {
-        $('.fixed.menu').transition('fade out')
-      }
-    })
-
-    // jQuery for page scrolling feature - requires jQuery Easing plugin
-    $('.ui.large.menu .left.menu a').bind('click', function() {
-      var $anchor = $(this)
-      $('html, body').stop().animate({
-          scrollTop: $($anchor.attr('href')).offset().top - 50
-      }, 1500, 'easeInOutExpo')
-      event.preventDefault()
-    })
-    console.log('lodaded!')
-  }
 
   logout(e) {
     e.preventDefault()
@@ -63,9 +27,7 @@ class HeaderNav extends Component {
     this.context.router.history.push('/login')
   }
 
-  render() {
-    
-    let isAuthenticated = false
+  render() {    
     let currentUser = null
 
     const authToken = localStorage.getItem('authToken')   
@@ -73,202 +35,153 @@ class HeaderNav extends Component {
     if (authToken) {
       const { user } = decode(authToken)
 
-      isAuthenticated = user ? true : false
       currentUser = user ? user : null
     }   
 
-    const UserLinks = () => {
+    const { getUnreadDirectMessagesCount } = this.props.data
+    let count = getUnreadDirectMessagesCount && getUnreadDirectMessagesCount.count 
 
-      const { getUnreadDirectMessagesCount } = this.props.data
-      let count = getUnreadDirectMessagesCount && getUnreadDirectMessagesCount.count 
+    // Count unread messages
+    const unreadMessagesCount = (T.translate("internal_navigation.unread_messages", {unread_messages_number: count}))
 
-      // Count unread messages
-      const unreadMessagesCount = (T.translate("internal_navigation.unread_messages", {unread_messages_number: count}))
+    const { getProjectTasksDataQuery: { getProjectTasksData }, getSaleTasksDataQuery: 
+      { getSaleTasksData}, getInvoiceTasksDataQuery: { getInvoiceTasksData} } = this.props
 
-      const { getProjectTasksDataQuery: { getProjectTasksData }, getSaleTasksDataQuery: 
-        { getSaleTasksData}, getInvoiceTasksDataQuery: { getInvoiceTasksData} } = this.props
+    // Count notificaitions
+    let countNotifications = 0
 
-      // Count notificaitions
-      let countNotifications = 0
+    const projectData = getProjectTasksData && getProjectTasksData.countStatus.map(item => {
+      if (item.status === 'delayed') {
+        countNotifications += item.count
 
-      const projectData = getProjectTasksData && getProjectTasksData.countStatus.map(item => {
-        if (item.status === 'delayed') {
-          countNotifications += item.count
+        return(<Dropdown.Item as='a' key={item.status}>
+          <Label color="red">{item.count} DELAYED</Label> 
+          Projects
+        </Dropdown.Item>)              
+      }
+      if (item.status === 'new') {
+        countNotifications += item.count
 
-          return(<Dropdown.Item as='a' key={item.status}>
-            <Label color="red">{item.count} DELAYED</Label> 
-            Projects
-          </Dropdown.Item>)              
-        }
-        if (item.status === 'new') {
-          countNotifications += item.count
+        return(<Dropdown.Item as='a' key={item.status}>
+          <Label color="blue">{item.count} NEW</Label> 
+          Projects
+        </Dropdown.Item>)
+      }
+    })
+  
+    const saleData = getSaleTasksData && getSaleTasksData.countStatus.map(item => {
+      if (item.status === 'delayed') {
+        countNotifications += item.count
 
-          return(<Dropdown.Item as='a' key={item.status}>
-            <Label color="blue">{item.count} NEW</Label> 
-            Projects
-          </Dropdown.Item>)
-        }
-      })
-    
-      const saleData = getSaleTasksData && getSaleTasksData.countStatus.map(item => {
-        if (item.status === 'delayed') {
-          countNotifications += item.count
+        return(<Dropdown.Item as='a' key={item.status}>
+          <Label color="red">{item.count} DELAYED</Label> 
+          Sales
+        </Dropdown.Item>)              
+      }
+      if (item.status === 'new') {
+        countNotifications += item.count
 
-          return(<Dropdown.Item as='a' key={item.status}>
-            <Label color="red">{item.count} DELAYED</Label> 
-            Sales
-          </Dropdown.Item>)              
-        }
-        if (item.status === 'new') {
-          countNotifications += item.count
+        return(<Dropdown.Item as='a' key={item.status}>
+          <Label color="blue">{item.count} NEW</Label> 
+          Sales
+        </Dropdown.Item>)
+      }
+    })
 
-          return(<Dropdown.Item as='a' key={item.status}>
-            <Label color="blue">{item.count} NEW</Label> 
-            Sales
-          </Dropdown.Item>)
-        }
-      })
+    const invoiceData = getInvoiceTasksData && getInvoiceTasksData.countStatus.map(item => {
+      if (item.status === 'delayed') {
+        countNotifications += item.count
 
-      const invoiceData = getInvoiceTasksData && getInvoiceTasksData.countStatus.map(item => {
-        if (item.status === 'delayed') {
-          countNotifications += item.count
+        return(<Dropdown.Item as='a' key={item.status}>
+          <Label color="red">{item.count} DELAYED</Label> 
+          Invoices
+        </Dropdown.Item>)              
+      }
+      if (item.status === 'pending') {
+        countNotifications += item.count
 
-          return(<Dropdown.Item as='a' key={item.status}>
-            <Label color="red">{item.count} DELAYED</Label> 
-            Invoices
-          </Dropdown.Item>)              
-        }
-        if (item.status === 'pending') {
-          countNotifications += item.count
+        return(<Dropdown.Item as='a' key={item.status}>
+          <Label color="orange">{item.count} PENDING</Label> 
+          Invoices
+        </Dropdown.Item>)
+      }
+    })
 
-          return(<Dropdown.Item as='a' key={item.status}>
-            <Label color="orange">{item.count} PENDING</Label> 
-            Invoices
-          </Dropdown.Item>)
-        }
-      })
+    const notifications = (T.translate("internal_navigation.notifications", {notifications_number: countNotifications}))
 
-      const notifications = (T.translate("internal_navigation.notifications", {notifications_number: countNotifications}))
-
-      return [
-        <nav key="nav" className="ui fixed menu">
-          <div className="left menu">
-            <a className="item anchor" onClick={this.props.toggleInnerSidebarVisibility}><i className="sidebar icon"></i></a>
-            <div className="logo item">
-              <Link to="/dashboard">
-                <img src={logoInverted} alt="logo-inverted" />
-              </Link>
-            </div>
-          </div>
-
-          <Menu.Menu position='right'>
-            <Dropdown pointing='top right' className='ui dropdown item' 
-              trigger={(<Icon name="alarm" className="mr-0" />)} icon={null} > 
-              <Dropdown.Menu>
-                <Dropdown.Item disabled>
-                  {notifications}
-                </Dropdown.Item>)
-                {projectData}
-                {saleData}
-                {invoiceData}
-              </Dropdown.Menu>
-            </Dropdown>  
-           
-            <Dropdown pointing='top right' className='ui dropdown item'
-              trigger={(<div>
-                <Icon name='mail' className="mr-0" />
-                {count !== 0 && <Label size="tiny" color="red" floating>{count !== 0 && count}</Label>}
-              </div>)} icon={null} >
-              <Dropdown.Menu>
-                <Dropdown.Item disabled>
-                  {unreadMessagesCount}             
-                </Dropdown.Item>
-                <Dropdown.Divider />
-                <Dropdown.Item
-                  onClick={() => {
-                    this.context.router.history.push('/conversations')
-                  }}
-                >              
-                  <strong className="turquoise">{T.translate("internal_navigation.see_all_messages")}</strong>
-                </Dropdown.Item>
-              </Dropdown.Menu>
-            </Dropdown>
-            
-            <Dropdown pointing='top right' className='ui dropdown item'
-              trigger={(
-                <span>
-                  <Image avatar src={avatarPlaceholderSmall} alt="avatar-placeholder-small" /> 
-                    {currentUser && currentUser.firstName}
-                </span>)} >
-              <Dropdown.Menu>
-                <Dropdown.Item>
-                  <Icon name="tasks"/>
-                  {T.translate("internal_navigation.tasks")}
-                  <div className="ui right floated blue label">{countNotifications}</div>
-                </Dropdown.Item>
-                <Dropdown.Item>
-                  <Link to='/settings'>
-                    <Icon name="settings" />
-                    {T.translate("internal_navigation.settings")}
-                  </Link>
-                </Dropdown.Item>
-                <Dropdown.Divider />
-                <Dropdown.Item as='a' onClick={this.logout.bind(this)}>
-                  <Icon name="sign out" />
-                  {T.translate("internal_navigation.sign_out")}
-                </Dropdown.Item>
-              </Dropdown.Menu>
-            </Dropdown>             
-          </Menu.Menu>
-        </nav>,
-
-        <Breadcrumb key="breadcrumb" />
-      ]
-    }
-
-    const GuestLinks = () => (
-      <div id="home" className="ui inverted vertical masthead center aligned segment">
-        <div className="ui container">
-          <div className="ui large secondary inverted pointing menu">
-            <a className="toc item" onClick={this.props.toggleOuterSidebarVisibility}>
-              <i className="sidebar icon"></i>
-            </a> 
-
-            <div className="left menu">
-              <ActiveLink activeOnlyWhenExact to="#home" label={T.translate("landing.home.header")} />
-              <ActiveLink activeOnlyWhenExact to="#features" label={T.translate("landing.features.header")} />
-              <ActiveLink activeOnlyWhenExact to="#clients" label={T.translate("landing.clients.header")} />
-              <ActiveLink activeOnlyWhenExact to="#testimonials" label={T.translate("landing.testmonial.header")} />
-              <ActiveLink activeOnlyWhenExact to="#pricing" label={T.translate("landing.pricing.header")} />
-              <ActiveLink activeOnlyWhenExact to="#contacts" label={T.translate("landing.contacts.header")} />
-            </div>
-
-            <div className="right item">                                                   
-              <Link className="ui inverted button"  to="/subdomain">{T.translate("log_in.log_in")}</Link> 
-              <a href={`${process.env.SERVER_PROTOCOL}${process.env.SERVER_HOST}/signup`} className="ui inverted button">{T.translate("sign_up.sign_up")}</a>     
-            </div>  
+    return [   
+      <nav key="headerNav" className="ui fixed menu">
+        <div className="left menu">
+          <a className="item anchor" onClick={this.props.toggleInnerSidebarVisibility}><i className="sidebar icon"></i></a>
+          <div className="logo item">
+            <Link to="/dashboard">
+              <img src={logoInverted} alt="logo-inverted" />
+            </Link>
           </div>
         </div>
 
-        <div className="ui text container">
-          <h1 className="ui inverted header">
-            {T.translate("landing.home.welcome")}&nbsp;
-            <div className="turquoise d-inline">{T.translate("internal_navigation.toolsio")}</div>
-          </h1>
-          <h3>{T.translate("landing.home.slogan")}</h3>
-          <a href={`${process.env.SERVER_PROTOCOL}${process.env.SERVER_HOST}/signup`} className="ui huge primary button">{T.translate("landing.home.get_started")}<i className="right arrow icon"></i></a>
-        </div>
-      </div>
-    )
-
-    return (
-      <header>    
-       
-        {/* Call links conditionally.  */}
-        { isAuthenticated ? <UserLinks /> : <GuestLinks /> } 
-
-      </header>
-    )
+        <Menu.Menu position='right'>
+          <Dropdown pointing='top right' className='ui dropdown item' 
+            trigger={(<Icon name="alarm" className="mr-0" />)} icon={null} > 
+            <Dropdown.Menu>
+              <Dropdown.Item disabled>
+                {notifications}
+              </Dropdown.Item>)
+              {projectData}
+              {saleData}
+              {invoiceData}
+            </Dropdown.Menu>
+          </Dropdown>  
+         
+          <Dropdown pointing='top right' className='ui dropdown item'
+            trigger={(<div>
+              <Icon name='mail' className="mr-0" />
+              {count !== 0 && <Label size="tiny" color="red" floating>{count !== 0 && count}</Label>}
+            </div>)} icon={null} >
+            <Dropdown.Menu>
+              <Dropdown.Item disabled>
+                {unreadMessagesCount}             
+              </Dropdown.Item>
+              <Dropdown.Divider />
+              <Dropdown.Item
+                onClick={() => {
+                  this.context.router.history.push('/conversations')
+                }}
+              >              
+                <strong className="turquoise">{T.translate("internal_navigation.see_all_messages")}</strong>
+              </Dropdown.Item>
+            </Dropdown.Menu>
+          </Dropdown>
+          
+          <Dropdown pointing='top right' className='ui dropdown item'
+            trigger={(
+              <span>
+                <Image avatar src={avatarPlaceholderSmall} alt="avatar-placeholder-small" /> 
+                  {currentUser && currentUser.firstName}
+              </span>)} >
+            <Dropdown.Menu>
+              <Dropdown.Item>
+                <Icon name="tasks"/>
+                {T.translate("internal_navigation.tasks")}
+                <div className="ui right floated blue label">{countNotifications}</div>
+              </Dropdown.Item>
+              <Dropdown.Item>
+                <Link to='/settings'>
+                  <Icon name="settings" />
+                  {T.translate("internal_navigation.settings")}
+                </Link>
+              </Dropdown.Item>
+              <Dropdown.Divider />
+              <Dropdown.Item as='a' onClick={this.logout.bind(this)}>
+                <Icon name="sign out" />
+                {T.translate("internal_navigation.sign_out")}
+              </Dropdown.Item>
+            </Dropdown.Menu>
+          </Dropdown>             
+        </Menu.Menu>
+      </nav>,
+      <Breadcrumb key="breadcrumb" />
+    ]
   }
 }
