+++ server/toolsio.template
@@ -1,9 +1,15 @@
 server {
+  
+  map $http_upgrade $connection_upgrade {
+    default upgrade;
+    '' close;
+  }
+
   location /subscriptions {
     proxy_pass http://web:8080/subscriptions;
     proxy_http_version 1.1;
     proxy_set_header Upgrade $http_upgrade;
-    proxy_set_header Connection "Upgrade";
+    proxy_set_header Connection $connection_upgrade;
   }
 
   location /graphql {
